refactor: split project-tools, extract executeSceneOp + BridgeManager + autoload-ini
Three structural extractions:

- Split monolithic `project-tools.ts` into runtime / autoload / project
  modules so each tool lives next to its peers.
- Extract `executeSceneOp` helper for the 15 headless-op handlers — one
  place for the execute + empty-stdout-check + try/catch pattern.
- Extract `BridgeManager` from `GodotRunner` (owns McpBridge artifact
  lifecycle: script copy, autoload entry, `.gdignore`, `.gitignore`
  augmentation) and `autoload-ini.ts` (single source of truth for
  `[autoload]` INI parse/add/remove/update).
