+++ tests/unit/handlers/autoload-handlers.test.ts
@@ -1,0 +1,230 @@
+import { describe, it, expect } from 'vitest';
+import { writeFileSync } from 'fs';
+import { join } from 'path';
+import {
+  handleListAutoloads,
+  handleAddAutoload,
+  handleRemoveAutoload,
+  handleUpdateAutoload,
+} from '../../../src/tools/autoload-tools.js';
+import { hasError, expectErrorMatching } from '../../helpers/assertions.js';
+import { fixtureProjectPath } from '../../helpers/fixture-paths.js';
+import { useTmpDirs } from '../../helpers/tmp.js';
+
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
+
+const tmp = useTmpDirs();
+
+/** Create a minimal tmp Godot project (project.godot only). */
+function makeTmpProject(): string {
+  return tmp.makeProject('mcp-test-');
+}
+
+/** Create a minimal project with one autoload registered. */
+function makeTmpProjectWithAutoload(name: string, path: string): string {
+  const dir = makeTmpProject();
+  const content = `config_version=5\n\n[autoload]\n${name}="*res://${path}"\n`;
+  writeFileSync(join(dir, 'project.godot'), content, 'utf8');
+  return dir;
+}
+
+// ---------------------------------------------------------------------------
+// handleListAutoloads
+// ---------------------------------------------------------------------------
+
+describe('handleListAutoloads', () => {
+  it('rejects missing projectPath', async () => {
+    const result = await handleListAutoloads({});
+    expectErrorMatching(result, /projectPath/i);
+  });
+
+  it('rejects projectPath containing ..', async () => {
+    const result = await handleListAutoloads({ projectPath: '../evil' });
+    expectErrorMatching(result, /invalid project path/i);
+  });
+
+  it('rejects nonexistent project directory', async () => {
+    const result = await handleListAutoloads({ projectPath: '/does/not/exist' });
+    expectErrorMatching(result, /not a valid godot project/i);
+  });
+
+  it('returns autoloads list for valid project', async () => {
+    const result = await handleListAutoloads({ projectPath: fixtureProjectPath });
+    expect(hasError(result)).toBe(false);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// handleAddAutoload
+// ---------------------------------------------------------------------------
+
+describe('handleAddAutoload', () => {
+  it('rejects missing projectPath', async () => {
+    const result = await handleAddAutoload({
+      autoloadName: 'MyManager',
+      autoloadPath: 'autoload/my.gd',
+    });
+    expectErrorMatching(result, /projectPath/i);
+  });
+
+  it('rejects projectPath containing ..', async () => {
+    const result = await handleAddAutoload({
+      projectPath: '../evil',
+      autoloadName: 'MyManager',
+      autoloadPath: 'autoload/my.gd',
+    });
+    expectErrorMatching(result, /invalid project path/i);
+  });
+
+  it('rejects nonexistent project', async () => {
+    const result = await handleAddAutoload({
+      projectPath: '/ghost',
+      autoloadName: 'MyManager',
+      autoloadPath: 'autoload/my.gd',
+    });
+    expectErrorMatching(result, /not a valid godot project/i);
+  });
+
+  it('rejects missing autoloadName', async () => {
+    const result = await handleAddAutoload({
+      projectPath: fixtureProjectPath,
+      autoloadPath: 'autoload/my.gd',
+    });
+    expect(hasError(result)).toBe(true);
+  });
+
+  it('rejects missing autoloadPath', async () => {
+    const result = await handleAddAutoload({
+      projectPath: fixtureProjectPath,
+      autoloadName: 'MyManager',
+    });
+    expect(hasError(result)).toBe(true);
+  });
+
+  it('rejects autoloadPath containing ..', async () => {
+    const result = await handleAddAutoload({
+      projectPath: fixtureProjectPath,
+      autoloadName: 'MyManager',
+      autoloadPath: '../outside.gd',
+    });
+    expect(hasError(result)).toBe(true);
+  });
+
+  it('registers autoload in a fresh tmp project', async () => {
+    const dir = makeTmpProject();
+    const result = await handleAddAutoload({
+      projectPath: dir,
+      autoloadName: 'TestManager',
+      autoloadPath: 'scripts/test.gd',
+    });
+    expect(hasError(result)).toBe(false);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// handleRemoveAutoload
+// ---------------------------------------------------------------------------
+
+describe('handleRemoveAutoload', () => {
+  it('rejects missing projectPath', async () => {
+    const result = await handleRemoveAutoload({ autoloadName: 'MyManager' });
+    expectErrorMatching(result, /projectPath/i);
+  });
+
+  it('rejects projectPath containing ..', async () => {
+    const result = await handleRemoveAutoload({
+      projectPath: '../evil',
+      autoloadName: 'MyManager',
+    });
+    expectErrorMatching(result, /invalid project path/i);
+  });
+
+  it('rejects nonexistent project', async () => {
+    const result = await handleRemoveAutoload({
+      projectPath: '/ghost',
+      autoloadName: 'MyManager',
+    });
+    expectErrorMatching(result, /not a valid godot project/i);
+  });
+
+  it('rejects missing autoloadName', async () => {
+    const result = await handleRemoveAutoload({ projectPath: fixtureProjectPath });
+    expect(hasError(result)).toBe(true);
+  });
+
+  it('returns isError when named autoload does not exist', async () => {
+    const result = await handleRemoveAutoload({
+      projectPath: fixtureProjectPath,
+      autoloadName: 'NonExistentAutoload',
+    });
+    expect(hasError(result)).toBe(true);
+  });
+
+  it('removes an existing autoload in a tmp project', async () => {
+    const dir = makeTmpProjectWithAutoload('TestManager', 'scripts/test.gd');
+    const result = await handleRemoveAutoload({ projectPath: dir, autoloadName: 'TestManager' });
+    expect(hasError(result)).toBe(false);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// handleUpdateAutoload
+// ---------------------------------------------------------------------------
+
+describe('handleUpdateAutoload', () => {
+  it('rejects missing projectPath', async () => {
+    const result = await handleUpdateAutoload({ autoloadName: 'MyManager' });
+    expectErrorMatching(result, /projectPath/i);
+  });
+
+  it('rejects projectPath containing ..', async () => {
+    const result = await handleUpdateAutoload({
+      projectPath: '../evil',
+      autoloadName: 'MyManager',
+    });
+    expectErrorMatching(result, /invalid project path/i);
+  });
+
+  it('rejects nonexistent project', async () => {
+    const result = await handleUpdateAutoload({
+      projectPath: '/ghost',
+      autoloadName: 'MyManager',
+    });
+    expectErrorMatching(result, /not a valid godot project/i);
+  });
+
+  it('rejects missing autoloadName', async () => {
+    const result = await handleUpdateAutoload({ projectPath: fixtureProjectPath });
+    expect(hasError(result)).toBe(true);
+  });
+
+  it('rejects autoloadPath containing ..', async () => {
+    const result = await handleUpdateAutoload({
+      projectPath: fixtureProjectPath,
+      autoloadName: 'MyManager',
+      autoloadPath: '../escape.gd',
+    });
+    expect(hasError(result)).toBe(true);
+  });
+
+  it('returns isError when named autoload does not exist', async () => {
+    const result = await handleUpdateAutoload({
+      projectPath: fixtureProjectPath,
+      autoloadName: 'NonExistentAutoload',
+      autoloadPath: 'scripts/new.gd',
+    });
+    expect(hasError(result)).toBe(true);
+  });
+
+  it('updates an existing autoload in a tmp project', async () => {
+    const dir = makeTmpProjectWithAutoload('TestManager', 'scripts/old.gd');
+    const result = await handleUpdateAutoload({
+      projectPath: dir,
+      autoloadName: 'TestManager',
+      autoloadPath: 'scripts/new.gd',
+    });
+    expect(hasError(result)).toBe(false);
+  });
+});
